@@ -8,17 +8,17 @@
     "node_modules"
   ],
   "dependencies": {
-    "@angular/common": "6.1.0",
-    "@angular/compiler": "6.1.0",
-    "@angular/core": "6.1.0",
-    "@angular/forms": "6.1.0",
-    "@angular/platform-browser": "6.1.0",
-    "@angular/platform-browser-dynamic": "6.1.0",
-    "@angular/router": "6.1.0",
-    "@fortawesome/angular-fontawesome": "0.2.0",
+    "@angular/common": "7.0.0",
+    "@angular/compiler": "7.0.0",
+    "@angular/core": "7.0.0",
+    "@angular/forms": "7.0.0",
+    "@angular/platform-browser": "7.0.0",
+    "@angular/platform-browser-dynamic": "7.0.0",
+    "@angular/router": "7.0.0",
+    "@fortawesome/angular-fontawesome": "0.2.1",
     "@fortawesome/fontawesome-svg-core": "1.2.4",
     "@fortawesome/free-solid-svg-icons": "5.3.1",
-    "@ng-bootstrap/ng-bootstrap": "3.0.0",
+    "@ng-bootstrap/ng-bootstrap": "4.0.0",
     "bootstrap": "4.1.3",
     "core-js": "2.5.7",
     "moment": "2.22.2",
@@ -71,23 +71,21 @@
     "mini-css-extract-plugin": "0.4.2",
     "moment-locales-webpack-plugin": "1.0.5",
     "optimize-css-assets-webpack-plugin": "5.0.1",
-    "prettier": "1.11.1",
+    "prettier": "1.14.3",
     "protractor": "5.4.0",
     "proxy-middleware": "0.15.0",
-    "raw-loader": "0.5.1",
     "rimraf": "2.6.1",
     "simple-progress-webpack-plugin": "1.1.2",
     "style-loader": "0.20.3",
-    "tapable": "1.0.0",
     "terser-webpack-plugin": "1.0.0",
     "thread-loader": "1.1.5",
     "to-string-loader": "1.1.5",
     "ts-node": "5.0.1",
-    "ts-loader": "4.0.1",
-    "tslint": "5.9.1",
-    "tslint-config-prettier": "1.9.0",
+    "ts-loader": "4.5.0",
+    "tslint": "5.11.0",
+    "tslint-config-prettier": "1.15.0",
     "tslint-loader": "3.6.0",
-    "typescript": "2.7.2",
+    "typescript": "3.1.3",
     "postcss-loader": "2.1.1",
     "xml2js": "0.4.19",
     "webpack": "4.8.0",
