@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { HttpResponse } from '@angular/common/http';
 import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot, Routes } from '@angular/router';
 import { UserRouteAccessService } from 'app/core';
-import { of } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
 import { BankAccount } from 'app/shared/model/bank-account.model';
 import { BankAccountService } from './bank-account.service';
 import { BankAccountComponent } from './bank-account.component';
@@ -16,10 +16,13 @@ import { IBankAccount } from 'app/shared/model/bank-account.model';
 export class BankAccountResolve implements Resolve<IBankAccount> {
     constructor(private service: BankAccountService) {}
 
-    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
+    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<BankAccount> {
         const id = route.params['id'] ? route.params['id'] : null;
         if (id) {
-            return this.service.find(id).pipe(map((bankAccount: HttpResponse<BankAccount>) => bankAccount.body));
+            return this.service.find(id).pipe(
+                filter((response: HttpResponse<BankAccount>) => response.ok),
+                map((bankAccount: HttpResponse<BankAccount>) => bankAccount.body)
+            );
         }
         return of(new BankAccount());
     }
