@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { HttpResponse } from '@angular/common/http';
 import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot, Routes } from '@angular/router';
 import { UserRouteAccessService } from 'app/core';
-import { of } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
+import { filter, map } from 'rxjs/operators';
 import { Operation } from 'app/shared/model/operation.model';
 import { OperationService } from './operation.service';
 import { OperationComponent } from './operation.component';
@@ -16,10 +16,13 @@ import { IOperation } from 'app/shared/model/operation.model';
 export class OperationResolve implements Resolve<IOperation> {
     constructor(private service: OperationService) {}
 
-    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
+    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Operation> {
         const id = route.params['id'] ? route.params['id'] : null;
         if (id) {
-            return this.service.find(id).pipe(map((operation: HttpResponse<Operation>) => operation.body));
+            return this.service.find(id).pipe(
+                filter((response: HttpResponse<Operation>) => response.ok),
+                map((operation: HttpResponse<Operation>) => operation.body)
+            );
         }
         return of(new Operation());
     }
