@@ -123,7 +123,7 @@ public class UserService {
         return newUser;
     }
     private boolean removeNonActivatedUser(User existingUser){
-        if(existingUser.getActivated()) {
+        if (existingUser.getActivated()) {
              return false;
         }
         userRepository.delete(existingUser);
