@@ -1,6 +1,5 @@
 import { Route } from '@angular/router';
 
-import { UserRouteAccessService } from 'app/core';
 import { ActivateComponent } from './activate.component';
 
 export const activateRoute: Route = {
@@ -9,6 +8,5 @@ export const activateRoute: Route = {
     data: {
         authorities: [],
         pageTitle: 'activate.title'
-    },
-    canActivate: [UserRouteAccessService]
+    }
 };
