@@ -36,10 +36,10 @@ export class PasswordStrengthBarComponent {
         force += 2 * p.length + (p.length >= 10 ? 1 : 0);
         force += passedMatches * 10;
 
-        // penality (short password)
+        // penalty (short password)
         force = p.length <= 6 ? Math.min(force, 10) : force;
 
-        // penality (poor variety of characters)
+        // penalty (poor variety of characters)
         force = passedMatches === 1 ? Math.min(force, 10) : force;
         force = passedMatches === 2 ? Math.min(force, 20) : force;
         force = passedMatches === 3 ? Math.min(force, 40) : force;
