@@ -11,73 +11,62 @@ describe('Component Tests', () => {
     describe('ActivateComponent', () => {
         let comp: ActivateComponent;
 
-        beforeEach(
-            async(() => {
-                TestBed.configureTestingModule({
-                    imports: [JhipsterSampleApplicationTestModule],
-                    declarations: [ActivateComponent],
-                    providers: [
-                        {
-                            provide: ActivatedRoute,
-                            useValue: new MockActivatedRoute({ key: 'ABC123' })
-                        }
-                    ]
-                })
-                    .overrideTemplate(ActivateComponent, '')
-                    .compileComponents();
+        beforeEach(async(() => {
+            TestBed.configureTestingModule({
+                imports: [JhipsterSampleApplicationTestModule],
+                declarations: [ActivateComponent],
+                providers: [
+                    {
+                        provide: ActivatedRoute,
+                        useValue: new MockActivatedRoute({ key: 'ABC123' })
+                    }
+                ]
             })
-        );
+                .overrideTemplate(ActivateComponent, '')
+                .compileComponents();
+        }));
 
         beforeEach(() => {
             const fixture = TestBed.createComponent(ActivateComponent);
             comp = fixture.componentInstance;
         });
 
-        it(
-            'calls activate.get with the key from params',
-            inject(
-                [ActivateService],
-                fakeAsync((service: ActivateService) => {
-                    spyOn(service, 'get').and.returnValue(of());
+        it('calls activate.get with the key from params', inject(
+            [ActivateService],
+            fakeAsync((service: ActivateService) => {
+                spyOn(service, 'get').and.returnValue(of());
 
-                    comp.ngOnInit();
-                    tick();
+                comp.ngOnInit();
+                tick();
 
-                    expect(service.get).toHaveBeenCalledWith('ABC123');
-                })
-            )
-        );
+                expect(service.get).toHaveBeenCalledWith('ABC123');
+            })
+        ));
 
-        it(
-            'should set set success to OK upon successful activation',
-            inject(
-                [ActivateService],
-                fakeAsync((service: ActivateService) => {
-                    spyOn(service, 'get').and.returnValue(of({}));
+        it('should set set success to OK upon successful activation', inject(
+            [ActivateService],
+            fakeAsync((service: ActivateService) => {
+                spyOn(service, 'get').and.returnValue(of({}));
 
-                    comp.ngOnInit();
-                    tick();
+                comp.ngOnInit();
+                tick();
 
-                    expect(comp.error).toBe(null);
-                    expect(comp.success).toEqual('OK');
-                })
-            )
-        );
+                expect(comp.error).toBe(null);
+                expect(comp.success).toEqual('OK');
+            })
+        ));
 
-        it(
-            'should set set error to ERROR upon activation failure',
-            inject(
-                [ActivateService],
-                fakeAsync((service: ActivateService) => {
-                    spyOn(service, 'get').and.returnValue(throwError('ERROR'));
+        it('should set set error to ERROR upon activation failure', inject(
+            [ActivateService],
+            fakeAsync((service: ActivateService) => {
+                spyOn(service, 'get').and.returnValue(throwError('ERROR'));
 
-                    comp.ngOnInit();
-                    tick();
+                comp.ngOnInit();
+                tick();
 
-                    expect(comp.error).toBe('ERROR');
-                    expect(comp.success).toEqual(null);
-                })
-            )
-        );
+                expect(comp.error).toBe('ERROR');
+                expect(comp.success).toEqual(null);
+            })
+        ));
     });
 });
