@@ -52,24 +52,21 @@ describe('Component Tests', () => {
             expect(comp.resetAccount).toEqual({});
         });
 
-        it(
-            'sets focus after the view has been initialized',
-            inject([ElementRef], (elementRef: ElementRef) => {
-                const element = fixture.nativeElement;
-                const node = {
-                    focus() {}
-                };
+        it('sets focus after the view has been initialized', inject([ElementRef], (elementRef: ElementRef) => {
+            const element = fixture.nativeElement;
+            const node = {
+                focus() {}
+            };
 
-                elementRef.nativeElement = element;
-                spyOn(element, 'querySelector').and.returnValue(node);
-                spyOn(node, 'focus');
+            elementRef.nativeElement = element;
+            spyOn(element, 'querySelector').and.returnValue(node);
+            spyOn(node, 'focus');
 
-                comp.ngAfterViewInit();
+            comp.ngAfterViewInit();
 
-                expect(element.querySelector).toHaveBeenCalledWith('#password');
-                expect(node.focus).toHaveBeenCalled();
-            })
-        );
+            expect(element.querySelector).toHaveBeenCalledWith('#password');
+            expect(node.focus).toHaveBeenCalled();
+        }));
 
         it('should ensure the two passwords entered match', () => {
             comp.resetAccount.password = 'password';
@@ -80,49 +77,43 @@ describe('Component Tests', () => {
             expect(comp.doNotMatch).toEqual('ERROR');
         });
 
-        it(
-            'should update success to OK after resetting password',
-            inject(
-                [PasswordResetFinishService],
-                fakeAsync((service: PasswordResetFinishService) => {
-                    spyOn(service, 'save').and.returnValue(of({}));
+        it('should update success to OK after resetting password', inject(
+            [PasswordResetFinishService],
+            fakeAsync((service: PasswordResetFinishService) => {
+                spyOn(service, 'save').and.returnValue(of({}));
 
-                    comp.resetAccount.password = 'password';
-                    comp.confirmPassword = 'password';
+                comp.resetAccount.password = 'password';
+                comp.confirmPassword = 'password';
 
-                    comp.finishReset();
-                    tick();
+                comp.finishReset();
+                tick();
 
-                    expect(service.save).toHaveBeenCalledWith({
-                        key: 'XYZPDQ',
-                        newPassword: 'password'
-                    });
-                    expect(comp.success).toEqual('OK');
-                })
-            )
-        );
+                expect(service.save).toHaveBeenCalledWith({
+                    key: 'XYZPDQ',
+                    newPassword: 'password'
+                });
+                expect(comp.success).toEqual('OK');
+            })
+        ));
 
-        it(
-            'should notify of generic error',
-            inject(
-                [PasswordResetFinishService],
-                fakeAsync((service: PasswordResetFinishService) => {
-                    spyOn(service, 'save').and.returnValue(throwError('ERROR'));
+        it('should notify of generic error', inject(
+            [PasswordResetFinishService],
+            fakeAsync((service: PasswordResetFinishService) => {
+                spyOn(service, 'save').and.returnValue(throwError('ERROR'));
 
-                    comp.resetAccount.password = 'password';
-                    comp.confirmPassword = 'password';
+                comp.resetAccount.password = 'password';
+                comp.confirmPassword = 'password';
 
-                    comp.finishReset();
-                    tick();
+                comp.finishReset();
+                tick();
 
-                    expect(service.save).toHaveBeenCalledWith({
-                        key: 'XYZPDQ',
-                        newPassword: 'password'
-                    });
-                    expect(comp.success).toBeNull();
-                    expect(comp.error).toEqual('ERROR');
-                })
-            )
-        );
+                expect(service.save).toHaveBeenCalledWith({
+                    key: 'XYZPDQ',
+                    newPassword: 'password'
+                });
+                expect(comp.success).toBeNull();
+                expect(comp.error).toEqual('ERROR');
+            })
+        ));
     });
 });
