@@ -13,16 +13,14 @@ describe('Component Tests', () => {
         let fixture: ComponentFixture<RegisterComponent>;
         let comp: RegisterComponent;
 
-        beforeEach(
-            async(() => {
-                TestBed.configureTestingModule({
-                    imports: [JhipsterSampleApplicationTestModule],
-                    declarations: [RegisterComponent]
-                })
-                    .overrideTemplate(RegisterComponent, '')
-                    .compileComponents();
+        beforeEach(async(() => {
+            TestBed.configureTestingModule({
+                imports: [JhipsterSampleApplicationTestModule],
+                declarations: [RegisterComponent]
             })
-        );
+                .overrideTemplate(RegisterComponent, '')
+                .compileComponents();
+        }));
 
         beforeEach(() => {
             fixture = TestBed.createComponent(RegisterComponent);
@@ -39,97 +37,85 @@ describe('Component Tests', () => {
             expect(comp.doNotMatch).toEqual('ERROR');
         });
 
-        it(
-            'should update success to OK after creating an account',
-            inject(
-                [Register, JhiLanguageService],
-                fakeAsync((service: Register, mockTranslate: MockLanguageService) => {
-                    spyOn(service, 'save').and.returnValue(of({}));
-                    comp.registerAccount.password = comp.confirmPassword = 'password';
+        it('should update success to OK after creating an account', inject(
+            [Register, JhiLanguageService],
+            fakeAsync((service: Register, mockTranslate: MockLanguageService) => {
+                spyOn(service, 'save').and.returnValue(of({}));
+                comp.registerAccount.password = comp.confirmPassword = 'password';
 
-                    comp.register();
-                    tick();
+                comp.register();
+                tick();
 
-                    expect(service.save).toHaveBeenCalledWith({
-                        password: 'password',
-                        langKey: 'en'
-                    });
-                    expect(comp.success).toEqual(true);
-                    expect(comp.registerAccount.langKey).toEqual('en');
-                    expect(mockTranslate.getCurrentSpy).toHaveBeenCalled();
-                    expect(comp.errorUserExists).toBeNull();
-                    expect(comp.errorEmailExists).toBeNull();
-                    expect(comp.error).toBeNull();
-                })
-            )
-        );
+                expect(service.save).toHaveBeenCalledWith({
+                    password: 'password',
+                    langKey: 'en'
+                });
+                expect(comp.success).toEqual(true);
+                expect(comp.registerAccount.langKey).toEqual('en');
+                expect(mockTranslate.getCurrentSpy).toHaveBeenCalled();
+                expect(comp.errorUserExists).toBeNull();
+                expect(comp.errorEmailExists).toBeNull();
+                expect(comp.error).toBeNull();
+            })
+        ));
 
-        it(
-            'should notify of user existence upon 400/login already in use',
-            inject(
-                [Register],
-                fakeAsync((service: Register) => {
-                    spyOn(service, 'save').and.returnValue(
-                        throwError({
-                            status: 400,
-                            error: { type: LOGIN_ALREADY_USED_TYPE }
-                        })
-                    );
-                    comp.registerAccount.password = comp.confirmPassword = 'password';
+        it('should notify of user existence upon 400/login already in use', inject(
+            [Register],
+            fakeAsync((service: Register) => {
+                spyOn(service, 'save').and.returnValue(
+                    throwError({
+                        status: 400,
+                        error: { type: LOGIN_ALREADY_USED_TYPE }
+                    })
+                );
+                comp.registerAccount.password = comp.confirmPassword = 'password';
 
-                    comp.register();
-                    tick();
+                comp.register();
+                tick();
 
-                    expect(comp.errorUserExists).toEqual('ERROR');
-                    expect(comp.errorEmailExists).toBeNull();
-                    expect(comp.error).toBeNull();
-                })
-            )
-        );
+                expect(comp.errorUserExists).toEqual('ERROR');
+                expect(comp.errorEmailExists).toBeNull();
+                expect(comp.error).toBeNull();
+            })
+        ));
 
-        it(
-            'should notify of email existence upon 400/email address already in use',
-            inject(
-                [Register],
-                fakeAsync((service: Register) => {
-                    spyOn(service, 'save').and.returnValue(
-                        throwError({
-                            status: 400,
-                            error: { type: EMAIL_ALREADY_USED_TYPE }
-                        })
-                    );
-                    comp.registerAccount.password = comp.confirmPassword = 'password';
+        it('should notify of email existence upon 400/email address already in use', inject(
+            [Register],
+            fakeAsync((service: Register) => {
+                spyOn(service, 'save').and.returnValue(
+                    throwError({
+                        status: 400,
+                        error: { type: EMAIL_ALREADY_USED_TYPE }
+                    })
+                );
+                comp.registerAccount.password = comp.confirmPassword = 'password';
 
-                    comp.register();
-                    tick();
+                comp.register();
+                tick();
 
-                    expect(comp.errorEmailExists).toEqual('ERROR');
-                    expect(comp.errorUserExists).toBeNull();
-                    expect(comp.error).toBeNull();
-                })
-            )
-        );
+                expect(comp.errorEmailExists).toEqual('ERROR');
+                expect(comp.errorUserExists).toBeNull();
+                expect(comp.error).toBeNull();
+            })
+        ));
 
-        it(
-            'should notify of generic error',
-            inject(
-                [Register],
-                fakeAsync((service: Register) => {
-                    spyOn(service, 'save').and.returnValue(
-                        throwError({
-                            status: 503
-                        })
-                    );
-                    comp.registerAccount.password = comp.confirmPassword = 'password';
+        it('should notify of generic error', inject(
+            [Register],
+            fakeAsync((service: Register) => {
+                spyOn(service, 'save').and.returnValue(
+                    throwError({
+                        status: 503
+                    })
+                );
+                comp.registerAccount.password = comp.confirmPassword = 'password';
 
-                    comp.register();
-                    tick();
+                comp.register();
+                tick();
 
-                    expect(comp.errorUserExists).toBeNull();
-                    expect(comp.errorEmailExists).toBeNull();
-                    expect(comp.error).toEqual('ERROR');
-                })
-            )
-        );
+                expect(comp.errorUserExists).toBeNull();
+                expect(comp.errorEmailExists).toBeNull();
+                expect(comp.error).toEqual('ERROR');
+            })
+        ));
     });
 });
