@@ -14,17 +14,15 @@ describe('Component Tests', () => {
         let fixture: ComponentFixture<JhiConfigurationComponent>;
         let service: JhiConfigurationService;
 
-        beforeEach(
-            async(() => {
-                TestBed.configureTestingModule({
-                    imports: [JhipsterSampleApplicationTestModule],
-                    declarations: [JhiConfigurationComponent],
-                    providers: [JhiConfigurationService]
-                })
-                    .overrideTemplate(JhiConfigurationComponent, '')
-                    .compileComponents();
+        beforeEach(async(() => {
+            TestBed.configureTestingModule({
+                imports: [JhipsterSampleApplicationTestModule],
+                declarations: [JhiConfigurationComponent],
+                providers: [JhiConfigurationService]
             })
-        );
+                .overrideTemplate(JhiConfigurationComponent, '')
+                .compileComponents();
+        }));
 
         beforeEach(() => {
             fixture = TestBed.createComponent(JhiConfigurationComponent);
